// __fixtures__/component/types.tsx
export type HabitType = {
  id: string;
  checked: boolean;
  text: string;
  deleted: boolean;
};

export type OptionalHabitType = {
  id?: string;
  checked?: boolean;
  text?: string;
  deleted?: boolean;
};

export type IHabitType = { [id: string]: HabitType };